requirements.txt
anyio==3.6.2
asyncpg==0.27.0
certifi==2022.12.7
click==8.1.3
databases==0.6.2
fastapi==0.91.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.4
python-dotenv==0.21.1
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.24.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
passlib[bcrypt]
alembic==1.9.3
